requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask_seeder
pyserial